modify: fixture활용 test코드 간결하ㅔㄱ
